services/gateway/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.2
prometheus-client==0.19.0
python-multipart==0.0.6
httpx==0.26.0
aiofiles==23.2.1
structlog==24.1.0
python-json-logger==2.0.7
tenacity==8.2.3
openai==1.54.0
python-dotenv==1.0.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
